+++ svtk/cxsv/__init__.py
@@ -1,2 +1,3 @@
-from .cpx_link import link_cpx, ComplexSV
+from .cpx_link import link_cpx
+from .complex_sv import ComplexSV
 from .rescan_single_enders import rescan_single_ender
